+++ src/types/helpers.ts
@@ -1,0 +1,7 @@
+export type RecursivePartial<T> = {
+	[P in keyof T]?: T[P] extends (infer U)[]
+		? RecursivePartial<U>[]
+		: T[P] extends object | undefined
+		? RecursivePartial<T[P]>
+		: T[P];
+};
