+++ src/types/tpsrc.ts
@@ -1,0 +1,10 @@
+import { TemplateAnswers, TemplateOptions } from './templates';
+
+export interface Tpsrc {
+	[templateName: string]: TpsrcTemplateConfig;
+}
+
+export interface TpsrcTemplateConfig {
+	opts: TemplateOptions;
+	answers: TemplateAnswers;
+}
